+++ app/src/main/java/ru/vsu/csf/bakebudget/models/OrderModel.kt
@@ -4,6 +4,6 @@ data class OrderModel(
     val id : Int,
     var status: Int,
     val product: ProductModel,
-    val finalPrice: Int,
+    val finalPrice: Double,
     val weight: Int,
 )
